# perf: Use a single mutex for both the stack and gc