--- a/c-api/src/lib.rs
+++ b/c-api/src/lib.rs
@@ -76,7 +76,7 @@ pub unsafe extern "C" fn glu_load_script(vm: &Thread,
 }
 
 pub extern "C" fn glu_call_function(thread: &Thread, arguments: VmIndex) -> Error {
-    let stack = thread.current_frame();
+    let stack = thread.get_stack();
     match thread.call_function(stack, arguments) {
         Ok(_) => Error::Ok,
         Err(_) => Error::Unknown,
@@ -84,11 +84,13 @@ pub extern "C" fn glu_call_function(thread: &Thread, arguments: VmIndex) -> Erro
 }
 
 pub extern "C" fn glu_len(vm: &Thread) -> usize {
-    vm.current_frame().len() as usize
+    let mut stack = vm.get_stack();
+    let stack = stack.current_frame();
+    stack.len() as usize
 }
 
 pub extern "C" fn glu_pop(vm: &Thread, n: usize) {
-    let mut stack = vm.current_frame();
+    let mut stack = vm.get_stack();
     for _ in 0..n {
         stack.pop();
     }
@@ -180,7 +182,8 @@ pub unsafe extern "C" fn glu_get_string(vm: &Thread,
                                         out: &mut *const u8,
                                         out_len: &mut usize)
                                         -> Error {
-    let stack = vm.current_frame();
+    let mut stack = vm.get_stack();
+    let stack = stack.current_frame();
     match stack.get_variants(index).and_then(|value| <&str>::from_value(vm, value)) {
         Some(value) => {
             *out = &*value.as_ptr();
@@ -206,7 +209,8 @@ pub extern "C" fn glu_get_light_userdata(vm: &Thread,
 fn get_value<T>(vm: &Thread, index: VmIndex, out: &mut T) -> Error
     where T: for<'vm> Getable<'vm>
 {
-    let stack = vm.current_frame();
+    let mut stack = vm.get_stack();
+    let stack = stack.current_frame();
     match stack.get_variants(index).and_then(|value| T::from_value(vm, value)) {
         Some(value) => {
             *out = value;