--- a/src/io.rs
+++ b/src/io.rs
@@ -15,7 +15,7 @@ use vm::stack::StackFrame;
 
 use vm::internal::Value;
 
-use super::Compiler;
+use super::{Compiler, Error};
 
 fn print_int(i: VmInt) -> IO<()> {
     print!("{}", i);
@@ -61,8 +61,8 @@ fn read_file<'vm>(file: WithVM<'vm, &GluonFile>, count: usize) -> IO<Array<'vm, 
         match file.read(&mut *buffer) {
             Ok(bytes_read) => {
                 let value = {
-                    let stack = vm.get_stack();
-                    match vm.alloc(&stack, &buffer[..bytes_read]) {
+                    let mut context = vm.context();
+                    match context.alloc(&buffer[..bytes_read]) {
                         Ok(value) => value,
                         Err(err) => return IO::Exception(format!("{}", err)),
                     }
@@ -119,35 +119,34 @@ fn read_line() -> IO<String> {
 
 /// IO a -> (String -> IO a) -> IO a
 fn catch_io(vm: &Thread) -> Status {
-    let mut stack = vm.get_stack();
-    let frame_level = stack.get_frames().len();
-    let action = StackFrame::current(&mut stack)[0];
-    stack.push(action);
-    0.push(vm, &mut stack).unwrap();
-    match vm.call_function(stack, 1) {
+    let mut context = vm.context();
+    let frame_level = context.stack.get_frames().len();
+    let action = StackFrame::current(&mut context.stack)[0];
+    context.stack.push(action);
+    0.push(vm, &mut context).unwrap();
+    match vm.call_function(context, 1) {
         Ok(_) => Status::Ok,
         Err(err) => {
-            let mut stack = vm.get_stack();
+            let mut context = vm.context();
             {
-                let mut stack = StackFrame::current(&mut stack);
+                let mut stack = StackFrame::current(&mut context.stack);
                 while stack.stack.get_frames().len() > frame_level {
-                    match stack.exit_scope() {
-                        Ok(new_stack) => stack = new_stack,
-                        Err(_) => return Status::Error,
+                    if let Err(_) =stack.exit_scope() {
+                        return Status::Error;
                     }
                 }
                 let callback = stack[1];
                 stack.push(callback);
             }
             let fmt = format!("{}", err);
-            let _ = fmt.push(vm, &mut stack);
-            0.push(vm, &mut stack).unwrap();
-            match vm.call_function(stack, 2) {
+            let _ = fmt.push(vm, &mut context);
+            0.push(vm, &mut context).unwrap();
+            match vm.call_function(context, 2) {
                 Ok(_) => Status::Ok,
                 Err(err) => {
-                    stack = vm.get_stack();
+                    context = vm.context();
                     let fmt = format!("{}", err);
-                    let _ = fmt.push(vm, &mut stack);
+                    let _ = fmt.push(vm, &mut context);
                     Status::Error
                 }
             }
@@ -155,47 +154,38 @@ fn catch_io(vm: &Thread) -> Status {
     }
 }
 
+fn clear_frames(err: Error, frame_level: usize, mut stack: StackFrame) -> IO<String> {
+    let trace = stack.stacktrace(frame_level);
+    let fmt = format!("{}\n{}", err, trace);
+    while stack.stack.get_frames().len() > frame_level {
+        if let Err(_) = stack.exit_scope() {
+            return IO::Exception(fmt);
+        }
+    }
+    IO::Exception(fmt)
+}
+
 fn run_expr(expr: WithVM<RootStr>) -> IO<String> {
     let WithVM { vm, value: expr } = expr;
-    let frame_level = vm.get_stack().get_frames().len();
+    let frame_level = vm.context().stack.get_frames().len();
     let run_result = Compiler::new().run_expr::<Generic<A>>(vm, "<top>", &expr);
-    let mut stack = vm.get_stack();
-    let mut stack = StackFrame::current(&mut stack);
+    let mut context = vm.context();
+    let stack = StackFrame::current(&mut context.stack);
     match run_result {
         Ok((value, typ)) => IO::Value(format!("{:?} : {}", value.0, typ)),
-        Err(err) => {
-            let trace = stack.stacktrace(frame_level);
-            let fmt = format!("{}\n{}", err, trace);
-            while stack.stack.get_frames().len() > frame_level {
-                match stack.exit_scope() {
-                    Ok(new_stack) => stack = new_stack,
-                    Err(_) => return IO::Exception(fmt),
-                }
-            }
-            IO::Exception(fmt)
-        }
+        Err(err) => clear_frames(err, frame_level, stack),
     }
 }
 
 fn load_script(name: WithVM<RootStr>, expr: RootStr) -> IO<String> {
     let WithVM { vm, value: name } = name;
-    let frame_level = vm.get_stack().get_frames().len();
+    let frame_level = vm.context().stack.get_frames().len();
     let run_result = Compiler::new().load_script(vm, &name[..], &expr);
-    let mut stack = vm.get_stack();
-    let mut stack = StackFrame::current(&mut stack);
+    let mut context = vm.context();
+    let stack = StackFrame::current(&mut context.stack);
     match run_result {
         Ok(()) => IO::Value(format!("Loaded {}", &name[..])),
-        Err(err) => {
-            let trace = stack.stacktrace(frame_level);
-            let fmt = format!("{}\n{}", err, trace);
-            while stack.stack.get_frames().len() > frame_level {
-                match stack.exit_scope() {
-                    Ok(new_stack) => stack = new_stack,
-                    Err(_) => return IO::Exception(fmt),
-                }
-            }
-            IO::Exception(fmt)
-        }
+        Err(err) => clear_frames(err, frame_level, stack),
     }
 }