--- a/test-nightly.sh
+++ b/test-nightly.sh
@@ -3,4 +3,4 @@ cargo test -p gluon_base --features test &&
     cargo test -p gluon_check --features test &&
     cargo test -p gluon_vm --features test &&
     cargo test --features "test nightly" &&
-    (cd c-api && cargo test --features test)
+    (cd c-api && cargo test --features "test nightly")