… tests/vm.rs
… vm/src/channel.rs
… vm/src/lazy.rs
… vm/src/primitives.rs
… vm/src/stack.rs
… vm/src/thread.rs