This gives a noticeable performance improvement for code which frequently enters and leaves the vm itself as only one mutex need to be locked. It does complicate some of the surrounding code slightly but this will be a necessary change regardless since the now added `Contex` type can be extended with additional fields which would otherwise involve adding more locks.

Before
```
test factorial                    ... bench:     249,725 ns/iter (+/- 12,150)
test factorial_tail_call          ... bench:     242,908 ns/iter (+/- 13,685)
test gluon_rust_boundary_overhead ... bench:     396,684 ns/iter (+/- 34,313)
```

After
```
test factorial                    ... bench:     128,571 ns/iter (+/- 7,537)
test factorial_tail_call          ... bench:     125,914 ns/iter (+/- 6,374)
test gluon_rust_boundary_overhead ... bench:     237,173 ns/iter (+/- 11,306)
```