@Marwes 2016-08-31
I am actually not entirely sure where what locking I removed to speedup the factorial benchmarks so much...