@Marwes 2016-09-03
Going to go ahead and merge this after travis passes since I have already based work in #132 on this. It can always be prettied up later.